--- backend/SLLBackend/src/main/java/service/sllbackend/controller/AuthController.java
@@ -1,16 +1,0 @@
-package service.sllbackend.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-// TODO: move to /web/
-
-@Controller
-@RequestMapping("/")
-public class AuthController {
-    @GetMapping("/login")
-    public String login() {
-        return "login";
-    }
-}
