--- backend/SLLBackend/src/main/java/service/sllbackend/controller/AuthRestController.java
@@ -1,23 +1,0 @@
-package service.sllbackend.controller;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import service.sllbackend.service.UserAccountService;
-
-// TODO: move to /web/
-
-@RestController
-@RequestMapping("/api")
-public class AuthRestController {
-    private final UserAccountService userAccountService;
-
-    public AuthRestController(UserAccountService userAccountService) {
-        this.userAccountService = userAccountService;
-    }
-
-    @PostMapping("/add")
-    public void add(){
-        userAccountService.registerUser("admin", "admin");
-    }
-}
